Added logic to limit a game to 5 rounds & output score.
Logic was added to create a game of 5 rounds, with the winner being who ever had the
higher score at the end of 5 rounds. Ties count as a complete round of RPS.
If no one has a higher score, a tie is declared.

Also added styling to indicate the round count and the player and computer scores
so that the user may understand the progress of the game.

Updated verbiage to have clearer meaning.