Removed deprecated code & changed game to winner = first to 5 wins
Some code that relied on user text input for choice was removed because it's no longer
neccesary with the implementation of buttons.

decalreWin() was added to clean up the checking and declaration of a winner.

Changed the game to be first to 5 wins.

Removed the portion that updated how many rounds had been played. It was a distraciton.